Move stringset into internal and create new commands using a constructor
