+++ internal/stringset/stringset.go
@@ -1,7 +1,7 @@
-// Package drouter was taken from docker's distribution repo, under Apache-2.0 license:
+// stringset package was taken from docker's distribution repo, under Apache-2.0 license:
 // https://github.com/docker/distribution/blob/0d3efadf015/registry/auth/token/stringset.go
 // Thank you for your work! <3
-package drouter
+package stringset
 
 // StringSet is a useful type for looking up strings.
 type StringSet map[string]struct{}
